# Uniform distribution: bias and usize portability