## Bias

I noticed that the `zone` produced via signed extension is not a multiple of the `range`. This means that the implementations for 8- and 16-bit types (@pitdicker) were biased. I added a fix for this which does not appear to have any performance impact.

It's worth mentioning however that the bias is *tiny*: the biggest deviation from a multiple of the range I could find was 32579 (for range = 65355); since this is sampling from a `u32` the probability of generating a biased sample is thus ~7.6e-6. With 100 million samples this bias is still lost in the noise.

Note: the included uniformity test is unfinished because it's pretty useless. Included for example but it should probably just be deleted.

## Portability for isize/usize samples

As discussed in #805. Unfortunately this does have a performance hit:
```
# before:
test distr_uniform_isize                   ... bench:       1,628 ns/iter (+/- 198) = 4914 MB/s
test distr_uniform_usize16                 ... bench:       1,617 ns/iter (+/- 121) = 4947 MB/s
test distr_uniform_usize32                 ... bench:       1,612 ns/iter (+/- 146) = 4962 MB/s
test distr_uniform_usize64                 ... bench:       2,473 ns/iter (+/- 79) = 3234 MB/s
# after:
test distr_uniform_isize                   ... bench:       5,818 ns/iter (+/- 121) = 1375 MB/s
test distr_uniform_usize16                 ... bench:       1,721 ns/iter (+/- 21) = 4648 MB/s
test distr_uniform_usize32                 ... bench:       1,803 ns/iter (+/- 20) = 4437 MB/s
test distr_uniform_usize64                 ... bench:       2,426 ns/iter (+/- 15) = 3297 MB/s
```
It also results in quite a bit of redundant, ugly code. As such I'm not happy about adding it (though it would be nice to have).

---

Thoughts? @burdges @vks @pitdicker