@burdges 2019-05-29
It's just the usual hit form rejection sampling, yes?