@dhardy 2019-05-30
All of these use rejection sampling, which makes the cost variable (especially when the required range is within a factor of 2 of the type's range). I'm not sure why the `isize` benchmark here suffers so much.