@vks 2019-05-31
> I'm not sure why the isize benchmark here suffers so much; the only additional cost should be a bit more branching.

Maybe it is branch misprediction? You could try running the benchmarks separately with `perf stat`.

I think we should have the bias corrections for sure. I'm not so sure about the portability improvements, there is a lot of code duplication (that could be reduced with macros). I would prefer not to promise value-stability for different platforms for platform-dependent types, and state this in the documentation. The tests could be fixed accordingly.

I think for sampling `usize` everyone expects that this is platform-dependent, but for `rand::seq` it is less intuitive. We could still consider implementing your fix there instead, or just mention the caveat in the documentation.