@dhardy 2019-06-01
Good point @vks that this would be better done within `seq` code. I count *eight* uses of `gen_range` and one of `Uniform<usize>` within the alias-method weighted index implementation. As such, implementing the suggestion is not trivial but not too hard.

Compatibility: the changes to `weighted::AliasMethod` are breaking if (a) you use exhaustive match on `weighted::WeightedError` or (b) you use `AliasMethod` with more than `u32::MAX` elements (talk about using gigabytes of memory and non-scalable algorithms!).

Performance: there are some minor wins and losses; nothing too significant I think:
```
# before:
running 19 tests
test misc_sample_indices_100_of_1G           ... bench:       2,562 ns/iter (+/- 122)
test misc_sample_indices_100_of_1M           ... bench:       2,496 ns/iter (+/- 182)
test misc_sample_indices_100_of_1k           ... bench:         549 ns/iter (+/- 23)
test misc_sample_indices_10_of_1k            ... bench:          89 ns/iter (+/- 4)
test misc_sample_indices_1_of_1k             ... bench:          26 ns/iter (+/- 1)
test misc_sample_indices_200_of_1G           ... bench:       5,048 ns/iter (+/- 271)
test misc_sample_indices_400_of_1G           ... bench:       9,010 ns/iter (+/- 228)
test misc_sample_indices_600_of_1G           ... bench:      13,227 ns/iter (+/- 966)
test seq_iter_choose_from_1000               ... bench:       3,341 ns/iter (+/- 46) = 2394 MB/s
test seq_iter_choose_multiple_10_of_100      ... bench:         931 ns/iter (+/- 142)
test seq_iter_choose_multiple_fill_10_of_100 ... bench:         873 ns/iter (+/- 42)
test seq_iter_unhinted_choose_from_1000      ... bench:       4,850 ns/iter (+/- 234)
test seq_iter_window_hinted_choose_from_1000 ... bench:       1,618 ns/iter (+/- 50)
test seq_shuffle_100                         ... bench:         837 ns/iter (+/- 31)
test seq_slice_choose_1_of_1000              ... bench:       3,336 ns/iter (+/- 211) = 2398 MB/s
test seq_slice_choose_multiple_10_of_100     ... bench:         156 ns/iter (+/- 16)
test seq_slice_choose_multiple_1_of_1000     ... bench:          33 ns/iter (+/- 2)
test seq_slice_choose_multiple_90_of_100     ... bench:         961 ns/iter (+/- 102)
test seq_slice_choose_multiple_950_of_1000   ... bench:       9,164 ns/iter (+/- 377)

running 8 tests
test distr_weighted_alias_method_f64       ... bench:      10,569 ns/iter (+/- 479) = 756 MB/s
test distr_weighted_alias_method_i8        ... bench:       9,906 ns/iter (+/- 603) = 807 MB/s
test distr_weighted_alias_method_large_set ... bench:      10,802 ns/iter (+/- 510) = 740 MB/s
test distr_weighted_alias_method_u32       ... bench:       9,898 ns/iter (+/- 915) = 808 MB/s
test distr_weighted_f64                    ... bench:       8,544 ns/iter (+/- 365) = 936 MB/s
test distr_weighted_i8                     ... bench:      11,145 ns/iter (+/- 931) = 717 MB/s
test distr_weighted_large_set              ... bench:      64,942 ns/iter (+/- 2,191) = 123 MB/s
test distr_weighted_u32                    ... bench:      10,906 ns/iter (+/- 414) = 733 MB/s

# after:
running 19 tests
test misc_sample_indices_100_of_1G           ... bench:       2,851 ns/iter (+/- 263)
test misc_sample_indices_100_of_1M           ... bench:       2,787 ns/iter (+/- 33)
test misc_sample_indices_100_of_1k           ... bench:         543 ns/iter (+/- 26)
test misc_sample_indices_10_of_1k            ... bench:          91 ns/iter (+/- 3)
test misc_sample_indices_1_of_1k             ... bench:          26 ns/iter (+/- 0)
test misc_sample_indices_200_of_1G           ... bench:       4,076 ns/iter (+/- 264)
test misc_sample_indices_400_of_1G           ... bench:       8,293 ns/iter (+/- 225)
test misc_sample_indices_600_of_1G           ... bench:      12,066 ns/iter (+/- 363)
test seq_iter_choose_from_1000               ... bench:       3,768 ns/iter (+/- 73) = 2123 MB/s
test seq_iter_choose_multiple_10_of_100      ... bench:         890 ns/iter (+/- 22)
test seq_iter_choose_multiple_fill_10_of_100 ... bench:         882 ns/iter (+/- 32)
test seq_iter_unhinted_choose_from_1000      ... bench:       4,842 ns/iter (+/- 183)
test seq_iter_window_hinted_choose_from_1000 ... bench:       1,816 ns/iter (+/- 41)
test seq_shuffle_100                         ... bench:         941 ns/iter (+/- 123)
test seq_slice_choose_1_of_1000              ... bench:       3,884 ns/iter (+/- 509) = 2059 MB/s
test seq_slice_choose_multiple_10_of_100     ... bench:         178 ns/iter (+/- 70)
test seq_slice_choose_multiple_1_of_1000     ... bench:          33 ns/iter (+/- 4)
test seq_slice_choose_multiple_90_of_100     ... bench:         973 ns/iter (+/- 69)
test seq_slice_choose_multiple_950_of_1000   ... bench:       9,152 ns/iter (+/- 377)

running 8 tests
test distr_weighted_alias_method_f64       ... bench:      10,665 ns/iter (+/- 175) = 750 MB/s
test distr_weighted_alias_method_i8        ... bench:       9,513 ns/iter (+/- 213) = 840 MB/s
test distr_weighted_alias_method_large_set ... bench:      10,378 ns/iter (+/- 537) = 770 MB/s
test distr_weighted_alias_method_u32       ... bench:       9,581 ns/iter (+/- 234) = 834 MB/s
test distr_weighted_f64                    ... bench:       8,508 ns/iter (+/- 215) = 940 MB/s
test distr_weighted_i8                     ... bench:      10,747 ns/iter (+/- 183) = 744 MB/s
test distr_weighted_large_set              ... bench:      65,314 ns/iter (+/- 3,106) = 122 MB/s
test distr_weighted_u32                    ... bench:      10,897 ns/iter (+/- 235) = 734 MB/s
```