@vks 2019-06-03
* Uniform distributions for SIMD types are currently broken.
* Do we want to document value stability among 32- and 64-bit platforms for `rand::seq`, or is it preferable to leave it unspecified?