@dhardy 2019-06-03
I dropped the uniformity test which was responsible for most of the failures and appears useless.