@vks 2019-06-03
The remaining failures will be fixed by #813, so I think this can be merged.