@dhardy 2019-06-03
Rebased on master; hopefully it passes this time